Refatorado o arquivo baseData.h e baseData.cpp
