--- gpPanel/mathplot.cpp
@@ -1,49 +1,0 @@
-
-
-//-----------------------------------------------------------------------------
-// mpLayer
-//-----------------------------------------------------------------------------
-
-
-//-----------------------------------------------------------------------------
-// mpInfoLayer
-//-----------------------------------------------------------------------------
-
-
-
-
-
-
-//-----------------------------------------------------------------------------
-// mpLayer implementations - functions
-//-----------------------------------------------------------------------------
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-//-----------------------------------------------------------------------------
-// mpLayer implementations - furniture (scales, ...)
-//-----------------------------------------------------------------------------
-
-
-
-
-
-
-
-
-
-
-
-
